+++ import_db2/mappings.py
@@ -210,14 +210,6 @@ CLIENT_DISCOUNT_PRICELIST = {
     600: '__setup__.pricelist_600',
 }
 
-CLIENT_PROMOTION_PRICELIST = {
-    db2_id: '__setup__.product_pricelist_supplier' for db2_id in
-    (100, 103, 104, 105, 106, 107, 108, 109, 110, 111, 112, 210, 211, 212,
-     213, 214, 301, 302, 303, 304, 308, 310, 311, 312, 314, 315, 316, 317,
-     318, 319, 320, 321, 322, 323, 401, 402, 403, 404, 405, 406, 407, 410,
-     411, 412, 600)
-}
-
 CLIENT_FISCAL_POSITION = {
   0: '__setup__.fiscal_position_nat',
   1: '__setup__.fiscal_position_nat',